Create Java Code
This allows a user to input five numbers. These numbers are stored in an array, and the program then recursively calculates the product of all five numbers. The recursive method multiplies each number in the array together, starting from the first element and continuing until the last element. Once the product is calculated, it is displayed to the user as the final result. This program demonstrates the use of recursion to solve a mathematical problem and showcases how recursive functions can be used to perform iterative calculations efficiently.